# src/main/java/com/arutech/mftracker/UserService/dto/PasswordResetRequestDto.java
package com.arutech.mftracker.UserService.dto;


import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@AllArgsConstructor
@NoArgsConstructor
@Getter
@Setter
public class PasswordResetRequestDto {
    private String email;
}